Add improvements to the goals and dashboard section by adding new endpoints and refactoring the walletRecommendation
