+++ src/lib/international.ts
@@ -1,23 +1,40 @@
 import Firestore from '../firebase';
 import { IInternationalItem } from './interfaces';
+import { getDataById } from '../commons/request';
+import defaultSectors from '../const/defaultSectors.json';
+import { uniqueArray } from '../builders/arrays';
 
-export const internationalSectors = [
-  'Stocks',
-  'Reits',
-  'Technology ETF',
-  'Gaming ETF',
-  'Medical ETF',
-  'Dividends ETF',
-  'Growth ETF',
-  'S&P ETF',
-  'Treasury',
-];
+// export const internationalSectors = [
+//   'Stocks',
+//   'Reits',
+//   'Technology ETF',
+//   'Gaming ETF',
+//   'Medical ETF',
+//   'Dividends ETF',
+//   'Growth ETF',
+//   'S&P ETF',
+//   'Treasury',
+// ];
 
-export const getInternationalSectors = async () => {
+export const getInternationalSectors = async (uid: string) => {
+  const userSectorsInternational = await getDataById(
+    'user:sectors:bonds',
+    'userSectorsInternational',
+    uid,
+    true,
+  );
+
+  const defaultInternationalSectors = defaultSectors.international;
+
   return {
     status: 200,
     message: 'Data Retrieved',
-    items: internationalSectors,
+    items: {
+      international: uniqueArray([
+        ...(userSectorsInternational.items?.values || []),
+        ...defaultInternationalSectors,
+      ]),
+    },
   };
 };
