+++ src/lib/user/walletRecommendation/const.ts
@@ -1,0 +1,7 @@
+export const assetNameMap = {
+  reits: 'fundos imobiliários',
+  stocks: 'ações',
+  bonds: 'renda fixa',
+  international: 'internacional',
+  crypto: 'crypto',
+};
