+++ src/services/assets.ts
@@ -1,5 +1,6 @@
 import * as assets from '../lib/assets';
 import Koa from 'koa';
+
 export const getAllSectors = async (ctx: Koa.Context) => {
   const { uid } = ctx.state.user;
   ctx.body = await assets.getAllSectors(uid);
