+++ src/services/international.ts
@@ -1,7 +1,9 @@
 import * as international from '../lib/international';
 import Koa from 'koa';
+
 export const getInternationalSectors = async (ctx: Koa.Context) => {
-  ctx.body = await international.getInternationalSectors();
+  const { uid } = ctx.state.user;
+  ctx.body = await international.getInternationalSectors(uid);
 };
 
 export const getInternationalAssets = async (ctx: Koa.Context) => {
